LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-history-entry/state-after-navigate-restore-expected.txt



PASS entry.getState() is properly restored after cross-document back
